src/app/components/modal/modal.component.spec.ts
import { TestBed } from '@angular/core/testing';
import { RouterTestingModule } from '@angular/router/testing';
import { ModalComponent } from './modal.component';

describe('AppComponent', () => {
  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [RouterTestingModule],
      declarations: [ModalComponent],
    }).compileComponents();
  });

  it('should create the app', () => {
    const fixture = TestBed.createComponent(ModalComponent);
    const app = fixture.componentInstance;
    expect(app).toBeTruthy();
  });

  it(`should have as title 'peachtree'`, () => {
    const fixture = TestBed.createComponent(ModalComponent);
    const app = fixture.componentInstance;
    expect(app.title).toEqual('peachtree');
  });

  it('should render title', () => {
    const fixture = TestBed.createComponent(ModalComponent);
    fixture.detectChanges();
    const compiled = fixture.nativeElement;
    expect(compiled.querySelector('.content span').textContent).toContain(
      'peachtree app is running!'
    );
  });
});
